data send and receive in API
