+++ src/index.js
@@ -1,17 +1,18 @@
-import "./Style/style.css";
-import { fetchData, setScore } from "./module/get_setdata.js";
-import DispalyLeaderData from "./module/DispalyLeaderGame.js";
+import './Style/style.css';
+import setScore from './module/Setdata.js';
+import DispalyLeaderData from './module/DispalyLeaderGame.js';
+
 DispalyLeaderData();
-const refBtn = document.querySelector(".btnSubmit");
-const formValue = document.querySelector("form");
-refBtn.addEventListener("click", async () => {
+const refBtn = document.querySelector('.btnSubmit');
+const formValue = document.querySelector('form');
+refBtn.addEventListener('click', async () => {
   DispalyLeaderData();
 });
-formValue.addEventListener("submit", async (event) => {
+formValue.addEventListener('submit', async (event) => {
   event.preventDefault();
 
-  const userName = document.querySelector(".userName").value;
-  const userScore = document.querySelector(".textScore").value;
+  const userName = document.querySelector('.userName').value;
+  const userScore = document.querySelector('.textScore').value;
 
   formValue.reset();
   await setScore(userName, userScore);
