+++ src/module/fetchdata.js
@@ -1,0 +1,14 @@
+import url from './gameAPI.js';
+
+const fetchData = async () => {
+  const response = await fetch(url, {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  });
+  const data = await response.json();
+  return data.result;
+};
+
+export default fetchData;
