+++ src/module/gameAPI.js
@@ -1,0 +1,2 @@
+const url = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/Zl4d7IVkemOTTVg2fUdz/scores';
+export default url;
