--- src/module/gamedata.js
@@ -1,3 +1,0 @@
-const url =
-  "https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/Zl4d7IVkemOTTVg2fUdz/scores";
-export default url;
